% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breached_site.R
\name{breached_site}
\alias{breached_site}
\title{Get a specific breached site, based in breach name (not domain)}
\usage{
breached_site(name, ...)
}
\arguments{
\item{name}{Search for a specific named breach}

\item{...}{Optional passthrough to HIBP_headers()}
}
\value{
Data.frame containing breach details
}
\description{
Get a specific breached site, based in breach name (not domain)
}
\examples{
breached_site("Adobe")
}
