% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GETcontent}
\alias{GETcontent}
\title{Used to extract content from a GET request and collapse to a data.frame}
\usage{
GETcontent(URLS, headers, ...)
}
\arguments{
\item{URLS}{urls}

\item{headers}{headers as built by HIBP_headers()}

\item{...}{Pass through arguments}
}
\value{
data.frame
}
\description{
It calls a helper function that retries up to 6 times
}
