% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIBP_headers.R
\name{HIBP_headers}
\alias{HIBP_headers}
\title{Construct headers for a HIBP request.
[Optional] Change the agent to make the request bespoke to you.}
\usage{
HIBP_headers(agent = "HIBPwned R pkg")
}
\arguments{
\item{agent}{The user agent the API call should run under}
}
\value{
headers
}
\description{
Construct headers for a HIBP request.
[Optional] Change the agent to make the request bespoke to you.
}
\examples{
HIBP_headers()
}

