% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_breaches.R
\name{account_breaches}
\alias{account_breaches}
\title{Search for data breaches associated with one or more accounts.}
\usage{
account_breaches(accounts, truncate = FALSE, domain = NULL,
  as_list = TRUE, ...)
}
\arguments{
\item{accounts}{A character vector of email addresses or account names}

\item{truncate}{Whether responses should be trimmed to just name of
breached site}

\item{domain}{Search a specific domain}

\item{as_list}{whether to return results as a list (default, TRUE) or a data.frame}

\item{...}{Optional passthrough to HIBP_headers()}
}
\value{
List of data.frames containing results
}
\description{
Search for data breaches associated with one or more accounts.
}
\examples{
account_breaches(c("steff.sullivan@gmail.com","stephanie.g.locke@gmail.com"))
}
