% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breached_sites.R
\name{breached_sites}
\alias{breached_sites}
\title{Get all (nonsensitive) breached sites in HIBP}
\usage{
breached_sites(domain = NULL, ...)
}
\arguments{
\item{domain}{Search a specific domain}

\item{...}{Optional passthrough to HIBP_headers()}
}
\value{
Data.frame containing breach details
}
\description{
Get all (nonsensitive) breached sites in HIBP
}
\examples{
breached_sites()
}

