% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GETcontent}
\alias{GETcontent}
\title{Used to extract content from a GET request and collapse to a data.frame}
\usage{
GETcontent(...)
}
\arguments{
\item{...}{Pass through arguments}
}
\value{
data.frame
}
\description{
Used to extract content from a GET request and collapse to a data.frame
}
