% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastes.R
\name{pastes}
\alias{pastes}
\title{Search for cases where an email address has been included in
a paste}
\usage{
pastes(emails, ...)
}
\arguments{
\item{emails}{A character vector of email addresses}

\item{...}{Optional passthrough to HIBP_headers()}
}
\value{
List of data.frames containing results
}
\description{
Search for cases where an email address has been included in
a paste
}
\examples{
pastes(c("steff.sullivan@gmail.com","stephanie.g.locke@gmail.com"))
}

